src/mysite/requirements.txt
Django==3.1.6
django-bootstrap4==2.3.1
django-allauth==0.44.0
mysqlclient==2.0.1
uwsgi==2.0.17